+++ lua/src/lzio.c
@@ -19,7 +19,7 @@
 
 
 int luaZ_fill (ZIO *z) {
-  size_t size;
+  unsigned int size;
   lua_State *L = z->L;
   const char *buff;
   lua_unlock(L);
@@ -55,9 +55,9 @@ void luaZ_init (lua_State *L, ZIO *z, lua_Reader reader, void *data) {
 
 
 /* --------------------------------------------------------------- read --- */
-size_t luaZ_read (ZIO *z, void *b, size_t n) {
+unsigned int luaZ_read (ZIO *z, void *b, unsigned int n) {
   while (n) {
-    size_t m;
+    unsigned int m;
     if (luaZ_lookahead(z) == EOZ)
       return n;  /* return number of missing bytes */
     m = (n <= z->n) ? n : z->n;  /* min. between n and z->n */
@@ -71,7 +71,7 @@ size_t luaZ_read (ZIO *z, void *b, size_t n) {
 }
 
 /* ------------------------------------------------------------------------ */
-char *luaZ_openspace (lua_State *L, Mbuffer *buff, size_t n) {
+char *luaZ_openspace (lua_State *L, Mbuffer *buff, unsigned int n) {
   if (n > buff->buffsize) {
     if (n < LUA_MINBUFFER) n = LUA_MINBUFFER;
     luaZ_resizebuffer(L, buff, n);
