+++ lua/src/lzio.h
@@ -23,8 +23,8 @@ typedef struct Zio ZIO;
 
 typedef struct Mbuffer {
   char *buffer;
-  size_t n;
-  size_t buffsize;
+  unsigned int n;
+  unsigned int buffsize;
 } Mbuffer;
 
 #define luaZ_initbuffer(L, buff) ((buff)->buffer = NULL, (buff)->buffsize = 0)
@@ -43,10 +43,10 @@ typedef struct Mbuffer {
 #define luaZ_freebuffer(L, buff)	luaZ_resizebuffer(L, buff, 0)
 
 
-LUAI_FUNC char *luaZ_openspace (lua_State *L, Mbuffer *buff, size_t n);
+LUAI_FUNC char *luaZ_openspace (lua_State *L, Mbuffer *buff, unsigned int n);
 LUAI_FUNC void luaZ_init (lua_State *L, ZIO *z, lua_Reader reader,
                                         void *data);
-LUAI_FUNC size_t luaZ_read (ZIO* z, void* b, size_t n);	/* read next n bytes */
+LUAI_FUNC unsigned int luaZ_read (ZIO* z, void* b, unsigned int n);	/* read next n bytes */
 LUAI_FUNC int luaZ_lookahead (ZIO *z);
 
 
@@ -54,7 +54,7 @@ LUAI_FUNC int luaZ_lookahead (ZIO *z);
 /* --------- Private Part ------------------ */
 
 struct Zio {
-  size_t n;			/* bytes still unread */
+  unsigned int n;			/* bytes still unread */
   const char *p;		/* current position in buffer */
   lua_Reader reader;
   void* data;			/* additional data */
